# gateways/api/openapi2-run.yaml
swagger: "2.0"
info:
  title: sterndeck-4
  version: "1.0.0"
host: api-dev.sterndeck.com
x-google-endpoints:
  - name: "api-dev.sterndeck.com"
    allowCors: True
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/sterndeck-4"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "sterndeck-4"
  sterndeck_api_key:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://api-dev.sterndeck.com/api/v1/auth-validate"
    x-google-jwks_uri: "https://api-dev.sterndeck.com/api/v1/auth-jwks"
    x-google-audiences: "sterndeck"
paths:
  /api/v1/auth-validate:
    options:
      summary: CORS pre-flight
      operationId: auth-validate-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    get:
      summary: "auth-validate"
      operationId: auth-validate
      security:
        - NONE: []
      responses:
        200:
          description: "auth-validate"
      x-google-backend:
        address: "https://http-auth-validate-i5rcwtaovq-uc.a.run.app"
  /api/v1/auth-jwks:
    options:
      summary: CORS pre-flight
      operationId: auth-jwks-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    get:
      summary: "auth-jwks"
      operationId: auth-jwks
      security:
        - NONE: []
      responses:
        200:
          description: "auth-jwks"
      x-google-backend:
        address: "https://http-auth-jwks-i5rcwtaovq-uc.a.run.app"
  /api/v1/auth-generate:
    options:
      summary: CORS pre-flight
      operationId: auth-generate-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "auth-generate"
      operationId: auth-generate
      security:
        - firebase: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              user_uid:
                type: string
      responses:
        200:
          description: "auth-generate"
      x-google-backend:
        address: "https://http-auth-generate-i5rcwtaovq-uc.a.run.app"

  /api/v1/completion:
    options:
      summary: CORS pre-flight
      operationId: completion-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "completion"
      operationId: completion
      security:
        - firebase: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              model:
                type: string
              prompt:
                type: string
              user_uid:
                type: string
      responses:
        200:
          description: "completion"
      x-google-backend:
        address: "https://http-completion-i5rcwtaovq-uc.a.run.app"
  /api/v1/user-key:
    options:
      summary: CORS pre-flight
      operationId: user-key-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "user-key"
      operationId: user-key
      security:
        - firebase: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              user_uid:
                type: string
              secret:
                type: string
      responses:
        200:
          description: "user-key"
      x-google-backend:
        address: "https://http-user-secret-create-i5rcwtaovq-uc.a.run.app"
  /api/v1/fine-tunes-create:
    options:
      summary: CORS pre-flight
      operationId: fine-tunes-create-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "fine-tunes-create"
      operationId: fine-tunes-create
      security:
        - firebase: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              user_uid:
                type: string
              training_file_id:
                type: string
              path:
                type: string
      responses:
        200:
          description: "fine-tunes-create"
      x-google-backend:
        address: "https://http-fine-tunes-create-i5rcwtaovq-uc.a.run.app"
  /api/v1/fine-tunes-cancel:
    options:
      summary: CORS pre-flight
      operationId: fine-tunes-cancel-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "fine-tunes-cancel"
      operationId: fine-tunes-cancel
      security:
        - firebase: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              user_uid:
                type: string
              fine_tune_id:
                type: string
              path:
                type: string
      responses:
        200:
          description: "fine-tunes-cancel"
      x-google-backend:
        address: "https://http-fine-tunes-cancel-i5rcwtaovq-uc.a.run.app"
  /api/v1/upload:
    options:
      summary: CORS pre-flight
      operationId: upload-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "upload"
      operationId: upload
      security:
        - firebase: []
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: upload
          type: string
          required: true
        - in: formData
          name: model_id
          type: string
          required: true
        - in: formData
          name: user_uid
          type: string
          required: true
        - in: formData
          name: file_id
          type: string
          required: true
      responses:
        200:
          description: "upload"
      x-google-backend:
        address: "https://http-csv-upload-i5rcwtaovq-uc.a.run.app"
  /api/v1/models-feedback:
    options:
      summary: CORS pre-flight
      operationId: feedback-cors
      security:
        - NONE: []
      x-google-backend:
        address: "https://http-hello-i5rcwtaovq-uc.a.run.app"
      responses:
        "200":
          description: OK
    post:
      summary: "feedback"
      operationId: feedback
      security:
        - sterndeck_api_key: []
        - firebase: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              model:
                type: string
              prompt:
                type: string
              completion:
                type: string
      responses:
        200:
          description: "feedback"
      x-google-backend:
        address: "https://http-models-feedback-i5rcwtaovq-uc.a.run.app"